Move main controller to controllers/
