+++ tests/application/test_main.py
@@ -21,7 +21,7 @@ class AppTestCase(FlaskAppTestCase):
         app.config['TESTING'] = True
         self.app = app.test_client()
 
-    @patch('application.main.signed_in')
+    @patch('application.controllers.main.signed_in')
     def test_homepage_redirects_to_upload_when_signed_in(self, signed_in):
         signed_in.return_value = True
         response = self.app.get('/')
@@ -46,8 +46,8 @@ class AppTestCase(FlaskAppTestCase):
         assert_that(response.data, contains_string(
             'Hello, Dave'))
 
-    @patch('application.main.signed_in_no_access')
-    @patch('application.main.get_authorization_url')
+    @patch('application.controllers.main.signed_in_no_access')
+    @patch('application.controllers.main.get_authorization_url')
     def test_homepage_redirects_to_auth_url_when_possible_access(
             self,
             get_authorization_url_patch,
